main functional with frontend and backend is done

+++ api/src/middleware/auth-headers.js
@@ -3,28 +3,40 @@ const { ACCESS_TOKEN_SECRET } = require('../../env.js');
 const Sessions = require('../models/token.js')
 
 module.exports = function(req, res, next) {
-    if (req.headers.authorization) {
-        jwt.verify(req.headers.authorization, ACCESS_TOKEN_SECRET, (err, payload) => {
+    if (req.body.headers && req.body.headers.authorization) {
+        jwt.verify(req.body.headers.authorization, ACCESS_TOKEN_SECRET, (err, payload) => {
             if (err) {
-                findSessionByToken(req.headers.authorization).then((candidate, error) => {
+                findSessionByToken(req.body.headers.authorization).then((candidate, error) => {
                     if (!error) {
                         deleteSession(candidate)
                         res.json({status: 401, message: 'Access denied'})
+                        if (!req.body.headers.skip) {
+                            next()
+                        }
                     } else {
                         res.json({status: 401, message: 'Access denied'})
+                        if (!req.body.headers.skip) {
+                            next()
+                        }
                     }
                 })
             } else if (payload) {
-                findUser(payload).then((candidate, error) => {
-                    if (error) {
+                findUser(payload).then((candidate) => {
+                    if (!candidate) {
                         res.json({status: 401, message: 'Not auth'})
+                        next()
                     } else {
-                        res.json({status: 200, message: 'Auth OK'})
+                        if (!req.body.headers.skip) {
+                            res.json({status: 200, message: 'Auth OK'})
+                        }
+                        next()
                     }
                 })
             }
         })
     }
+
+  //  next()
 }
 
 async function findUser(payload) {
